developed API for menu items and refactor all model id from int to long

+++ src/main/java/com/bobby/artistweb/model/AboutMe.java
@@ -14,7 +14,7 @@ public class AboutMe {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id = 0;
+    private long id = 0;
 
     private String name ="";
     @Column(columnDefinition = "TEXT")
