+++ src/main/java/com/bobby/artistweb/model/Applications.java
@@ -17,7 +17,7 @@ public class Applications {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id = 0;
+    private long id = 0;
     private String appName = "users";
     private String appKey = "7dfb4cf67742cb0660305e56ef816c53fcec892cae7f6ee39b75f34e659d672c";
