+++ src/main/java/com/bobby/artistweb/repo/ContactMeRepo.java
@@ -3,5 +3,5 @@ package com.bobby.artistweb.repo;
 import com.bobby.artistweb.model.ContactMe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContactMeRepo extends JpaRepository<ContactMe, Integer> {
+public interface ContactMeRepo extends JpaRepository<ContactMe, Long> {
 }
