+++ src/main/java/com/bobby/artistweb/repo/LogoRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LogoRepo extends JpaRepository<Logo, Integer> {
+public interface LogoRepo extends JpaRepository<Logo, Long> {
 }
